+++ R/utils-methods.R
@@ -1,9 +1,33 @@
-#' Import Matrix Methods
+#' Import Matrix
 #'
 #' Get methods and classes from Matrix
 #'
 #' @import methods
 #' @import Matrix
+#' @importMethodsFrom Matrix colSums
+#' @importMethodsFrom Matrix t
+#' @importMethodsFrom Matrix colMeans
+#' @importMethodsFrom Matrix rowSums
+#' @importMethodsFrom Matrix rowMeans
+#' @importMethodsFrom Matrix %*%
+#' @importMethodsFrom Matrix as.matrix
+#' @importMethodsFrom Matrix as.array
+#' @importMethodsFrom Matrix as.logical
+#' @importMethodsFrom Matrix as.numeric
+#' @importMethodsFrom Matrix as.vector
+#' @importMethodsFrom Matrix diag
+#' @importMethodsFrom Matrix norm
+#' @importMethodsFrom Matrix tail
+#' @importMethodsFrom Matrix head
+#' @importMethodsFrom Matrix mean
+#' @importMethodsFrom Matrix dim
+#' @importMethodsFrom Matrix length
+#' @importMethodsFrom Matrix image
+#' @importMethodsFrom Matrix all
+#' @importMethodsFrom Matrix any
+#' @importMethodsFrom Matrix is.na
+#' @importMethodsFrom Matrix dimnames
+#' @importMethodsFrom Matrix [
 #' @importClassesFrom Matrix dgCMatrix dgTMatrix
 #' @keywords internal
 #' @export